Remove the obsolete Go 1.14 build tag
