+++ cipher_suite.go
@@ -76,8 +76,6 @@ const (
 	CipherSuiteKeyExchangeAlgorithmEcdhe CipherSuiteKeyExchangeAlgorithm = ciphersuite.KeyExchangeAlgorithmEcdhe
 )
 
-var _ = allCipherSuites() // Necessary until this function isn't only used by Go 1.14
-
 // CipherSuite is an interface that all DTLS CipherSuites must satisfy.
 type CipherSuite interface {
 	// String of CipherSuite, only used for logging
@@ -109,10 +107,13 @@ type CipherSuite interface {
 	Decrypt(h recordlayer.Header, in []byte) ([]byte, error)
 }
 
+// VersionDTLS12 is the DTLS version in the same style as VersionTLSXX from crypto/tls.
+const VersionDTLS12 = 0xfefd
+
 // CipherSuiteName provides the same functionality as tls.CipherSuiteName
 // that appeared first in Go 1.14.
 //
-// Our implementation differs slightly in that it takes in a CiperSuiteID,
+// Our implementation differs slightly in that it takes in a CipherSuiteID,
 // like the rest of our library, instead of a uint16 like crypto/tls.
 func CipherSuiteName(id CipherSuiteID) string {
 	suite := cipherSuiteForID(id, nil)
@@ -121,6 +122,37 @@ func CipherSuiteName(id CipherSuiteID) string {
 	}
 
 	return fmt.Sprintf("0x%04X", uint16(id))
+}
+
+// Convert from our cipherSuite interface to a tls.CipherSuite struct.
+func toTLSCipherSuite(c CipherSuite) *tls.CipherSuite {
+	return &tls.CipherSuite{
+		ID:                uint16(c.ID()),
+		Name:              c.String(),
+		SupportedVersions: []uint16{VersionDTLS12},
+		Insecure:          false,
+	}
+}
+
+// CipherSuites returns a list of cipher suites currently implemented by this
+// package, excluding those with security issues, which are returned by
+// InsecureCipherSuites.
+func CipherSuites() []*tls.CipherSuite {
+	suites := allCipherSuites()
+	res := make([]*tls.CipherSuite, len(suites))
+	for i, c := range suites {
+		res[i] = toTLSCipherSuite(c)
+	}
+
+	return res
+}
+
+// InsecureCipherSuites returns a list of cipher suites currently implemented by
+// this package and which have security issues.
+func InsecureCipherSuites() []*tls.CipherSuite {
+	var res []*tls.CipherSuite
+
+	return res
 }
 
 // Taken from https://www.iana.org/assignments/tls-parameters/tls-parameters.xml
